Add size and position validation for modal CUIWnds.
Refactor InGameMenu to override SizeMove() to calculate its own position
and size (instead of DoLayout()) and move the calculation of layout
parameters into helper functions.  NB InGameMenu::SizeMove() doesn't call
CUIWnd::SizeMove() because the size/position validation is redundant.

Add modal_begin() and _end() to GG::GUI to return const_iterators to the
private list of modal wnds in GG::GUI and change
HumanClientApp::HandleWindowResize() to use this to get pointers to modal
CUIWnds and call ValidatePosition() on them.

Adjust MapWnd::DoLayout() to validate the position of all the fleet
windows, not just the active one, since it's possible to have more than
one visible at a time.