Create RichText.
RichText is a control that generates child
controls based on the text content it is given.

It parses the content for tags and creates child
controls for tags that have been registered.
It creates them using factory objects provided
in the registration.

Text not inside tags will be considered plaintext
and will be treated as content of a PLAINTEXT_TAG.
By default this text is shown using TextBlock.