Bump required CMake version to 2.8.5
The updated version is the first one providing the `GNUInstallDirs`
module suggested in #333.  Also taking a look at the [CMake Life cycle]
considerations page it should be clear that distributions providing only
a 2.6.* version don't belong to our primary target group and we already
use modules, properties and commands, that were not part of CMake 2.6.

[CMake Life cycle]: https://cmake.org/Wiki/CMake_Life_Cycle_Considerations