Fix Layout::DoLayout() recursion error with SetMinSize().
In Layout::DoLayout() causes the following bug.
If the size when DoLayout() is called is smaller than the new min size,
then when SetMinSize() is called DoLayout() is called again.  When the
second invocation of DoLayout() completes, the first invocation
continues from where it called SetMinSize().  The result of the
completion of the first invocation of SizeMove() depends on whether the
(ul,lr) pair of references, reference a member variable that was
correctly adjusted by the second invocation of DoLayout().

The fix is in two parts.

First, use m_stop_resize_recursion to prevent the second invocation of
DoLayout() from completing.

Second, change the const Pt& ul and const Pt& lr to copies not
references.  Pt is small, a pair of doubles and the pointer reference
doesn't save space.   const implies that ul and lr are constant, but
they may not be due to the mutating function calls within DoLayout().