Remove hazardous variant of Font::TextExtent.
Font had a variant of TextExtent() that 'helpfully' internally called
ExpensiveParseFromTextToTextElements and DetermineLines, computing and
discarding text_elements each time they were called.  This variant ran
10 times slower than the variant with the pre-computed text elements.

When called during Render() of other tight loops this results in delays
long enough to be perceived by humans.

This commit removes this TextExtent() variant.

It results in more verbose code for hazardous operations of repeatedly
generating text elements, merely to find a text extent.  It will
prompt the programmer to move ExpensiveParseFromTextToTextElements out
of the loop.