Remove Control* ListBox::Row::operator[](n).
ListBox::Row::operator[](n) is a negative feature.

The problem is that when given an out of range value it creates and
returns a dangling pointer to a Control*.  Even if the calling function
handles that correctly locally, the entire state of the Row is now
inconsistent.

If operator[] is called on an empty Row, then subsequent calls to
empty() return false, and calls to size() return 1 when the Row has no
valid Controls and m_cells[0] = 0 a null pointer.

This commit removes operator[] and replaces the use of

(*row)[0]

in the code with

!row->empty() ? row->at(0) : 0

which is the same result.

In places where the code did not check for null, I've fixed that as
well.