Replace UI signal functors with lambdas
Some actions that are a response to an UI signal the action needed
context data to execute that was only available at action-to-signal
connect time.  With C++03 we used dedicated functor classes to store the
context data.  This moved the action implementation away from it's place
of usage and required a declaration for the functor class.  Replacing
the Functors with C++11 lambdas places the action implementation right
at the place of usage and removes the verbosity required for a regular
functor.