In DropDownList ModalListPicker capture mousewheel events from child ListBox.
Redirection of mouse events and keyboard events for modal windows are
handled by code in GUI.cpp that directs the keyboard events to the top
level modal window and the mouse events to the child list box.

This commit adds an event filter to capture the mouse wheel events from
the child list box.