Derive GG::Wnd from std::enable_shared_from_this
This is the a step in converting GG to use the std smart pointers to
manage windows.

Calling shared_from_this() within a constructor will throw bad_weak_ptr
because construction of its internal weak_ptr is incomplete.  This will
have the consequence that AttachChild() can not be called from within
either the parent or the child window's constructor.  Any constructors
that need to call AttachChild() need to be converted to factory
functions.